+++ react-hot-button-component/src/index.jsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import HotButton from './hot-button';
 
 class MainContainer extends React.Component {
   render() {
     return (
       <div>
+        <HotButton />
       </div>
     );
   }
